+++ site-modules/role/spec/classes/puppet_server_spec.rb
@@ -7,6 +7,7 @@ describe 'role::puppet_server' do
   it { is_expected.to compile.with_all_deps }
   it { is_expected.to contain_class('profile::baseline') }
   it { is_expected.to contain_class('profile::administration_tools') }
+  it { is_expected.to contain_class('profile::lifecycle_state') }
   it { is_expected.to contain_class('profile::application_state') }
   it { is_expected.to contain_class('profile::agent_service') }
   it { is_expected.to contain_class('profile::server_operations') }